backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
openai==1.10.0
google-generativeai==0.3.2
langchain==0.1.4
langchain-community==0.0.20
chromadb==0.4.22
sentence-transformers==2.3.1
schedule==1.2.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0


